+++ index.spec.ts
@@ -10,6 +10,7 @@ import {
 import * as m from "mithril/render/hyperscript";
 
 import middleware, {
+    parsePath,
     router,
 } from "./index";
 
@@ -99,5 +100,52 @@ describe(router.name, () => {
     });
     it("should have working defaut route", async () => {
         expect(await router(defs, "not-exists", "/abc/abc")).to.be.a("string");
+    });
+});
+
+describe(parsePath.name, () => {
+    it("should return string", () => {
+        expect(parsePath("")).to.be.a("string");
+    });
+    it("should parse params", () => {
+        let query;
+        let hash;
+
+        query = {};
+        hash = {};
+        expect(parsePath("", query, hash)).to.be.a("string");
+        expect(query).to.be.empty;
+        expect(hash).to.be.empty;
+
+        query = {};
+        hash = {};
+        expect(parsePath("/?test=test#test=test", query, hash)).to.be.a("string");
+        expect(query).to.be.have.property("test").equal("test");
+        expect(hash).to.be.have.property("test").equal("test");
+
+        query = {};
+        hash = {};
+        expect(parsePath("/#test=test", query)).to.be.a("string");
+        expect(query).to.be.have.property("test").equal("test");
+
+        query = {};
+        hash = {};
+        expect(parsePath("/?test=test", query, hash)).to.be.a("string");
+        expect(query).to.be.have.property("test").equal("test");
+        expect(hash).to.be.empty;
+
+        query = {};
+        hash = {};
+        expect(parsePath("/#test=test", query, hash)).to.be.a("string");
+        expect(query).to.be.empty;
+        expect(hash).to.be.have.property("test").equal("test");
+    });
+    it("should return pathname", () => {
+        expect(parsePath("")).to.be.equal("");
+        expect(parsePath("/")).to.be.equal("/");
+        expect(parsePath("/test")).to.be.equal("/test");
+        expect(parsePath("/test?abc")).to.be.equal("/test");
+        expect(parsePath("/test#abc")).to.be.equal("/test");
+        expect(parsePath("/test?abc#abc")).to.be.equal("/test");
     });
 });
