+++ index.ts
@@ -1,3 +1,6 @@
+// tslint:disable-next-line:no-reference
+/// <reference path="./types/mithril.d.ts" />
+
 import {
     RequestHandler,
 } from "express";
@@ -6,11 +9,36 @@ import {
     RouteDefs,
 } from "mithril";
 
-import {
-    parsePath,
-} from "./parse";
-
 import render from "./render";
+
+import * as parseQueryString from "mithril/querystring/parse";
+
+// code from mithril.js internal method
+export function parsePath(
+    path: string,
+    queryData?: { [_: string]: any },
+    hashData: typeof queryData = queryData,
+): string {
+    const queryIndex = path.indexOf("?");
+    const hashIndex = path.indexOf("#");
+    const pathEnd = queryIndex > -1 ? queryIndex : hashIndex > -1 ? hashIndex : path.length;
+    if ((queryIndex > -1) && (typeof queryData !== "undefined")) {
+        const queryEnd = hashIndex > -1 ? hashIndex : path.length;
+        const queryParams = parseQueryString(path.slice(queryIndex + 1, queryEnd));
+        // tslint:disable-next-line:forin
+        for (const key in queryParams) {
+            queryData[key] = queryParams[key];
+        }
+    }
+    if ((hashIndex > -1) && (typeof hashData !== "undefined")) {
+        const hashParams = parseQueryString(path.slice(hashIndex + 1));
+        // tslint:disable-next-line:forin
+        for (const key in hashParams) {
+            hashData[key] = hashParams[key];
+        }
+    }
+    return path.slice(0, pathEnd);
+}
 
 export async function router(
     routes: RouteDefs,
